Initial commit
Uses PyQt to create a minimalistic GUI and psutils to poll all processes every 5 seconds. If it's currently one of the set work hours of the user and a process that matches the filter is started, it KILLS the process.
Current shortcomings: Settings are not saved in between executions, processes are KILLED instead of the execution being prevented (not sure if that's possible). Ugly UI :)